Prevent full reindex in both indexeres when running via setup:upgrade, move common indexer methods to abstract indexer. This fixes also the emulation error.
