+++ Model/Indexer/Invalidate.php
@@ -45,7 +45,6 @@ use Nosto\Tagging\Model\Indexer\Dimensions\ModeSwitcherInterface;
 use Nosto\Tagging\Model\ResourceModel\Magento\Product\Collection as ProductCollection;
 use Nosto\Tagging\Model\ResourceModel\Magento\Product\CollectionFactory as ProductCollectionFactory;
 use Nosto\Tagging\Model\Service\Index as NostoServiceIndex;
-use Nosto\Tagging\Util\Indexer as IndexerUtil;
 use Nosto\Tagging\Logger\Logger as NostoLogger;
 use Magento\Indexer\Model\ProcessManager;
 use Nosto\Tagging\Model\Indexer\Dimensions\Invalidate\ModeSwitcher as InvalidateModeSwitcher;
@@ -74,9 +73,6 @@ class Invalidate extends AbstractIndexer
 
     /** @var InvalidateModeSwitcher */
     private $modeSwitcher;
-
-    /** @var InputInterface */
-    private $input;
 
     /**
      * Invalidate constructor.
@@ -107,53 +103,15 @@ class Invalidate extends AbstractIndexer
         $this->nostoHelperAccount = $nostoHelperAccount;
         $this->productCollectionFactory = $productCollectionFactory;
         $this->modeSwitcher = $modeSwitcher;
-        $this->input = $input;
         parent::__construct(
             $nostoHelperAccount,
             $nostoHelperScope,
             $logger,
             $dimensionProvider,
             $storeEmulation,
+            $input,
             $processManager
         );
-    }
-
-    /**
-     * @param int[] $ids
-     * @throws Exception
-     */
-    public function execute($ids)
-    {
-        $this->doWork($ids);
-    }
-
-    /**
-     * @inheritDoc
-     * @throws NostoException
-     */
-    public function executeFull()
-    {
-        if ($this->allowFullExecution() === true) {
-            $this->doWork();
-        }
-    }
-
-    /**
-     * @inheritDoc
-     * @throws Exception
-     */
-    public function executeList(array $ids)
-    {
-        $this->execute($ids);
-    }
-
-    /**
-     * @inheritDoc
-     * @throws Exception
-     */
-    public function executeRow($id)
-    {
-        $this->execute([$id]);
     }
 
     /**
@@ -212,13 +170,5 @@ class Invalidate extends AbstractIndexer
             $collection->addActiveAndVisibleFilter();
         }
         return $collection;
-    }
-
-    /**
-     * @return bool
-     */
-    public function allowFullExecution()
-    {
-        return IndexerUtil::isCalledFromSetupUpgrade($this->input) === false;
     }
 }
